I add util for looking up wallet nfts in format for user profile.
